backend/requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
pydantic==2.7.0
python-dotenv==1.0.1
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
openai==1.2.0
pytest==7.4.3 